[GCU] Enhance test_eth_interface (#19341)
What is the motivation for this PR?
When interface FEC is not configured (N/A or none). The GCU rollback fails and port become operational down due to incorrect FEC.
Enhance the testcase to set FEC back to N/A (none).

How did you do it?
Enhance the testcase to set FEC back to N/A (none).

How did you verify/test it?
Verified by run test on Arista-7050CX3 M1 testbed.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>