Fix invalid vars format in playbook for Ansible compatibility (#19366)
What is the motivation for this PR?
In an Ansible playbook, the vars field must be defined as a dictionary (i.e., a set of key-value pairs), not as a list prefixed with -. When upgrading to a newer version of Ansible, the parser may enforce stricter validation on playbook syntax, which can result in errors like the following:

TASK [vm_set : Kickstart VMs] **************************************************
ERROR! Vars in a Task must be specified as a dictionary.

The error appears to be in '/var/src/sonic-mgmt/ansible/roles/vm_set/tasks/kickstart_vm.yml': line 154, column 7, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

    vars:
      - ansible_connection: 'network_cli'
      ^ here
This PR fixes the issue by correcting the structure of the vars section in the playbook to use the proper dictionary format, ensuring compatibility with newer Ansible versions and preventing similar errors in the future.

How did you do it?
This PR fixes the issue by correcting the structure of the vars section in the playbook to use the proper dictionary format, ensuring compatibility with newer Ansible versions and preventing similar errors in the future.

How did you verify/test it?
We need to make sure that this change won't affect current test firstly -- test by pipeline itself. And then, we need to make sure that this change works in the new version -- test locally.

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>