Added UCD checksum recomputation for BFD responder (#18853)
What is the motivation for this PR?
BFD responder is used in testing VPP platform. Currently the responder does not send the correct UDP checksum for the packets sent to DUT so the packets are dropped in VPP.

How did you do it?
Reset the BFD responder packet payload checksum to trigger a UDP checksum recomputation.

How did you verify/test it?
Tested locally with sonic-mgmt.

Any platform specific information?
T1 and T1-lag Sonic VPP.

Signed-off-by: Yael Tzur <ytzur@nvidia.com>