Merged PR 13835: workaround for testbed with tornado fanout
workaround for testbed with tornado fanout: using c1c1 instead of c1 in the testbed name for tornado fanout to preserve the 200G port speed.

----
#### AI description  (iteration 1)
#### PR Classification
This PR implements a bug fix in testbed configuration for a tornado fanout setup.

#### PR Summary
The pull request refines the testbed setup by adjusting conditions for port speed configuration and updating the testbed identifier.
- `ansible/config_sonic_basedon_testbed.yml`: Modified the conditional statement to exclude testbeds with 'c1c1', ensuring the port speed fix is applied only under intended conditions.
- `ansible/testbed.yaml`: Updated the configuration name to correctly reflect the testbed identifier for a tornado fanout environment.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->