Update iptables to block BGP (TCP 179) traffic on eth0 (#18834)
Description of PR
Summary: This PR updates the iptables and ip6tables rules to block incoming BGP (TCP port 179) traffic on the eth0 interface. This change ensures that BGP sessions are only allowed on non-management interfaces.
Fixes: N/A

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202205
 202305
 202311
 202405
 202411
 202505
Approach
What is the motivation for this PR?
To support test updates in this PR: sonic-host-services#197.
Additionally, it ensures BGP port 179 is not exposed on the management interface (eth0).

How did you do it?
How did you verify/test it?
On t0-64 testbed

co-authorized by: jianquanye@microsoft.com

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>