Disable PR testing for 202012 branch (#19417)
What is the motivation for this PR?
Support for release 202012 has ended. This PR disables PR pipeline triggers for the 202012 branch.

How did you do it?
Set PR trigger to "none" in pipeline yaml file.

How did you verify/test it?
Verify that PR testing is not triggered for this PR.