Fix flakiness with poll mode tests when client has not connected yet or default route has not populated yet (#19316)
What is the motivation for this PR?
This PR does the following:

Uses netstat to ensure that there is an established TCP connection between the client and server which is a more reliable check instead of pid check. This new check will help ensure that client is connected and can receive all notifications.

Ensures that we are giving enough time for default route to populate in APPL_DB after bgp sessions have been restored. There are some situations where we grab 10 updates, but default route has not been populated yet, so we miss the update. We will let the query run for longer and check less frequently to ensure that we see the default route entries after restoring bgp sessions.

How did you do it?
Code change

How did you verify/test it?
202411 test

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>