[logrotate] Fix the issue that no logs will be consider as non-zero return in grep (#19481)
Description of PR
Summary:
Fixes #33258706
[logrotate] Fix the issue that no logs will be consider as non-zero return in grep

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement

Approach
What is the motivation for this PR?
Fix the issue that no logs will be consider as non-zero return in grep

How did you do it?
Ignore errors of grep return

How did you verify/test it?
Tested on physical testbeds and it works well

signed-off-by: jianquanye@microsoft.com
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>