Add topology support for NUT testbeds. (#19429)
What is the motivation for this PR?
When defining the testbeds, it turns to be many testbeds will share the same topology. Repeating this info in all the testbeds is not ideal.

How did you do it?
Extracted DUT template and TG template out of the testbed definition and make them a topology for NUT.

How did you verify/test it?
Yes, deployed the testbed locally and all passing.

./testbed-cli.sh -t testbed.nut.yaml gen-cfg nut-sn5640-stress-1 ixia ../../password.txt

Signed-off-by: yenlu <yenlu@cisco.com>