Adding L1 support for Network Under Test (NUT) testbed (#19410)
What is the motivation for this PR?
The L1 switch are frequently used in the IXIA testbeds, so we need to get L1 supported.

How did you do it?
Adding L1 support for Network Under Test (NUT) testbed.

How did you verify/test it?
Tested locally and testbed are properly deployed.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>