Add capability for the advanced reboot test to collect tcpdump on the server physical port
Get the tcpdump on the physical interface of the test server instead of the logical ptf interfaces if an option is explicitly set to True in the pytest command. This is to fix the issue that sometimes there could be random packet drop on the logical bridges or ptf.
Keep the original approch to capture on the ptf interfaces untouched as the default flow.