Drop_packets Support needed for dstmac equals src mac, reserved mac and multicast source mac
*Description of Issue*
Drop Packets Test: - Support needed for forwarding packets from FANOUT(SONIC OS with marvell-teralynx ASIC) when any of the following conditions are met:
  1. The destination MAC address is the same as the source MAC address.
  2. The destination MAC address is a reserved MAC address.
  3. The source MAC address is a multicast address.

Requirements:
  In the TL10 Fanout, we are observing that certain source mac equals destination mac invalid packets are being dropped in fanout hardware. While we require TL10 fanout to  accept and forward these packets instead of dropping them.

Implementation:
we have generated ACL(Access Control List) rules from CSV for each VLAN. ACL can identify the packet on the based on the VLAN id and redirect the packet to the egress.
Didn't allows the pipeline to drop the pkt.
  1. Drop Packets testcases generates the acl_rules.json as per the Fanout port breakout.
  2. Same ACL config is removed post the testcase completes.
  3. Drop Packets testcases need to skip falcon as fanout

Signed-off-by: Anshul Dubela <adubela@marvell.com>