Fix PFCWD Test port filtering
Selected test ports are not meant to include loops with their rx ports.
The existing selection process did not correctly apply this,
particularly in the case of only 2 test ports available.

Problem situation seen in test_pfcwd_multi_port. When there are only 2
test ports they are eachother's RX ports and so when storm is started
on one, SendVerifyTraffic could not determine if the other was working
correctly.