Add check to verify if portA key exists in the pick_ports method
In pick_ports we assume portA always exists and this causes indexing errors.
There may be cases where portA is not found, for example, when the test looks for a portchannel connected to downstream T1/LT2 VMs.
In t2_single_node_max the downstream links do not have portchannels, and hence no intf will satisfy this condition.