Calculating the egress rate based on the total packets received for lossy and lossless streams. (#19173)
Calculating the egress rate based on the total packets received for lossy and lossless streams.

Description of PR
Some tests were failing with error "Egress speed beyond tolerance range"
On debugging it was found that tests case queries the egress rate from ixia, which keeps fluctuating. Depending upon when test case took the stats, it can affect the test case result. So to avoid this if the test fails based on ixia egress rate stats one more check is introduced to calculate the egress rate based on the total packets received and time traffic was sent and then check if the Egress rate is within the tolerance limit.

Summary:
Fixes # https://migsonic.atlassian.net/browse/MIGSMSFT-1029

Type of change
[ x] Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202205
 202305
 202311
[x ] 202405
 202411
 202505
Approach
Test case was failing

How did you do it?
Added the another check to calculate the Egress rate base on total packages received and time traffic was transmitted.

How did you verify/test it?
Manually ran the test to check the logic. Asking Austin to test it in MSFT setup and we are not able to reproduce this issue anymore.

signed-off-by: jianquanye@microsoft.com

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>