Add options to by pass cache cleanup before running the tests. (#19549)
What is the motivation for this PR?
RunTest CLI will always clean up the cache before running the tests, which will take time when we have multiple DUTs in the testbed.

How did you do it?
This PR adds an option -w to skip the cache clear and save time for test preparation.

How did you verify/test it?
Finished locally.

After this change, local cache will be used directly.

Signed-off-by: Yael Tzur <ytzur@nvidia.com>