Add markdown lint config file. (#19566)
What is the motivation for this PR?
The markdown docs might contain many format issues today. And having a markdownlint could greatly help with the doc quality.

How did you do it?
This PR adds a markdown lint config file. But it is not enforced yet, it only works for the ones with markdownlint installed for suggestion.

How did you verify/test it?
Installed markdownlint-cli and ran the lint locally.

npm install -g markdownlint-cli

This config also works for vscode markdownlint extensions.