[test_container_checker] disable_memory_utilization  (#19604)
What is the motivation for this PR?
The cases failed during teardown caused by the memory alarm.
Based script log, it should be related to the timing, when memory utilization collected the memory, some dockers were not up, so the memory before the test is much lower than normal, and caused memory alarm.

How did you do it?
Try to wait all critical processes up and bgp up after tests to ensure the memory information is correct.
However, based on the test result, the waiting time is much longer than expected since this is the case for docker restart.
So plan to disable the memory checker in this case test_container_checker, since
1: This case is for docker check, only do the stop_service then check the docker status. memory checker is not necessary.
2: If we want to wait all process up and bgp up, it would take much extra time for this case.

How did you verify/test it?
Elastic

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>