Reduce the number of ifconfig calls made in add_topo (#19119)
What is the motivation for this PR?
To reduce the setup time for add_topo, particularly for topologies with a larger number of VMs.

How did you do it?
Reduced the number of ifconfig -a calls within VMTopology.init(...) from to 1.

How did you verify/test it?
Ran add_topo on a t1-isolated topology and confirmed that the setup time was reduced.

Any platform specific information?
This change reduces setup time for non-multi-DUT configurations only; it does not impact the setup time for multi-DUT configurations.

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>