Add deviation modification for cisco-8000. (#19598)
Description of PR
Summary:
pfcwd_basic tests have to be tuned on platform basis. Since they are timer-sensitive, and can fail in corner cases with the symptom as below:

>           pytest_assert(loss_rate <= max_loss_rate and loss_rate >= min_loss_rate,
                          'Loss rate of {} ({}) should be in [{}, {}]'.format(
                              data_flow_name_list[i], loss_rate, min_loss_rate, max_loss_rate))
E           Failed: Loss rate of Data Flow 1 (0.5811225531621866) should be in [0.7, 1]
In this PR, we are increasing the deviation for cisco-8000 to 0.5 instead of the original 0.3.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202205
 202305
 202311
 202405
 202411
 202505
Approach
What is the motivation for this PR?
Cisco-8000 runs keep hitting flaky test fails due to above signature.

How did you do it?
Updated the deviation for cisco-8000 to 0.5.

How did you verify/test it?
Run in progress for entire pfcwd_basic. Will update the PR once it is done.

Any platform specific information?
Specific to cisco-8000.

signed-off-by: jianquanye@microsoft.com
Signed-off-by: yenlu <yenlu@cisco.com>