Merged PR 14486: [nightly hawk] Fix query to surface accurate failures
Issue:
all acl failed consisently on master image, but it is not surfaced out.

Before add the condition `| where Result in (ResultList`) for `dataClean`:
The Total count is different with failed count, because there are 12 skipped results in dataClean.

| Build ID           | Full Case Name                                                           | Total Attempts | Failed Attempts | Min Attempt | Max Attempt |
|--------------------|---------------------------------------------------------------------------|----------------|------------------|--------------|--------------|
| 687f5d0a080f912bf07127fb | acl.test_acl.TestIncrementalAcl.test_l4_sport_range_match_forwarded | 24             | 12               | 0            | 2            |

We have to remove them from dataClean, otherwise some consistent or flaky failures will not be able to surface out.

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>