Fix PFCWD Test port filtering (#19530)
What is the motivation for this PR?
This bug was originally found in msft-202503 where the test failed with
AssertionError: Did not receive expected packet on any of ports [] for device .

How did you do it?
I noticed this occurred with only 2 test ports available, so I changed the selection process.
#11656

I found this PR which originally mentioned the test_pfcwd_multi_port and included a change for 2 ports. But the situation where there are only 2 test ports available is problematic as storming one will mean traffic can't be verified on the other during this test case since they are eachother's RX ports.

How did you verify/test it?
I re-ran the test to see that test_pfcwd_multi_port should have skipped in this situation. And also verified that when there were enough test ports available test_pfcwd_multi_port did not skip.

Signed-off-by: yenlu <yenlu@cisco.com>