Update tools for running test on NUT testbeds. (#19241)
What is the motivation for this PR?
With the support of network under test (NUT) testbed being added, we need to update the tools for running the tests.

How did you do it?
Since the only difference between these 2 testbeds is the testbed definition, and the ixia tests uses very minimum set fields, we can easily create some adapting code to make the existing tool supporting the new type of testbeds.

How did you verify/test it?
Running test locally.

Signed-off-by: yenlu <yenlu@cisco.com>