TestbedProcessing: Align makeLabYaml with makeLab for Python3 compatibility
With #11457, in Python3 sonic-mgmt docker environments, we are forced to
use YAML modules to load files and this has repercussions if inventory
file is in INI format.

If we are using TestbedProcessing.py to generate veos (which has ansible
inventory entries for test server) and lab (which has inventory entries
for dut, ptf, and fanout switch), we would be forced to use it with yaml
flag as with #11457, we can only open yaml files.

If you take a look at makeLab and makeLabYaml, there are so many
required fields that are not at all handled in current makeLabYaml
function in its current state. This patch attempts to fix these
differences and make the transition to Python3 seamless if we are using
TestbedProcessing.py to generate inventory files.

Signed-off-by: Mohan Yelugoti <ymd@arista.com>