[master][sonic-mgmt]Skip ipv6 cases under everflow/test_everflow_per_interface.py on on Arista-7260CX3 and Arista-7060CX (#19500)
What is the motivation for this PR?
Added further cases to skip as requested in 19308
The original issue that requires these cases to be skipped is being tracked in 627 and public issue 19096
How did you do it?
How did you verify/test it?
Any platform specific information?
Yes - Arista-7260CX3(TH2)

Signed-off-by: Yael Tzur <ytzur@nvidia.com>