[DHCPv6_Relay]: Ensuring Client Reachability
It is observed in some cases test --> "dhcp_relay/test_dhcpv6_relay.py::TestDhcpv6RelayWithMultipleVlan::test_dhcp_relay_default"
fails if the neighbor entry for client LLA is not present.

To make sure client address is resolved before we start the PTF test,
ping to a multicast address has been done so that we discover all the LLAs
in the segment.