Add nexthop threshold for arista-7050cx3-32s-c6s104 (#19770)
What is the motivation for this PR?
There was an assert failure in test_crm_available.py for this SKU because it defaults to a nexthop threshold of 256 which is higher than the threshold the chip actually supports.

How did you do it?
Added an entry under SKU_NEXTHOP_THRESHOLDS dictionary to set the nexthop threshold to 255 which is in line with other SKUs that have the same switch ASIC.

How did you verify/test it?
Ran the test with and without the change. Passing with the change.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>