[Test plan] Snappi-based BGP Convergence Test (#19564)
What is the motivation for this PR?
Currently, we don't have any test plan that uses tgen for measure the BGP convergence performance and data plane downtime, especially in multi-tier network topology.

How did you do it?
This PR adds the test plan that uses tgen for testing, which is based on NUT testbed.

How did you verify/test it?
Doc only.

Signed-off-by: yenlu <yenlu@cisco.com>