Update codeowners (#19488)
What is the motivation for this PR?
This change updated the .github/CODEOWNERS to add some owners and fixed some issues.

How did you do it?
Added owners for the .azure-pipelines folder and some other folders.
Removed owners no longer working on the SONiC project.
Sort the entries in alphabetic order.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>