[vm_topology] Fix pid=0 issue if ptf docker is not running (#19793)
In vm_topology.py code, lots of the code check if the PTF container pid is 'None'. For example:

```
    if self.pid is None
```

In case PTF container is stopped, the current code would store PTF container pid "0" in `self.pid`.

Consequently, some of the logic is incorrect. For example, remove-topo while PTF container is stopped, the unbind topology step would fail if the host device happens to have a "eth0" interface.

The fix is to check the Running states of PTF container. If the container is not running, explicitly set `self.pid` to `None`.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>