wait for the change to be applied after async patch requests in test_restapi_vxlan_ecmp (#19762)
Since the PATCH request we call in patch_config_vrouter_vrf_id_routes is async, we can't verify the result right after we receive a response.
Because of that, I've added a check of existing route via CLI before performing a GET request to get the list of routes from restapi.

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>