[test_voq_chassis_app_db_consistency.py] Ensure get_db_dump is not called on rebooting DUT (#19981)
Approach
What is the motivation for this PR?
test_voq_chassis_app_db_consisitency.py reboot_dut case on dT2 would fail with Timeout (62s) waiting for privilege escalation prompt: as it would try to perform get_db_dump on the rebooting dut. Not a problem for T2 as the DUT rebooting is always a frontend node and get_db_dump is always called on a supervisor node.

How did you do it?
Made a change so test would only check_db_consistency on a rebooting DUT if there are more than 1 DUT (basically if it's T2)

How did you verify/test it?
Ran the test, didn't see the same failure.

Signed-off-by: Javier Tan javiertan@microsoft.com
Signed-off-by: Yael Tzur <ytzur@nvidia.com>