[SNAPPI] pfc/test_pfc_pause_lossless_warm_reboot.py Failure: ValueError: invalid literal for int() with base 10: 'N/A' (#19817)
Created a assert with a wait_until and checked for 60s / every 5s until PFC counters are not N/A.

What is the motivation for this PR?
Test case errors out with this message:
ValueError: invalid literal for int() with base 10: 'N/A'
The reason was PFC counters after a reboot was not available right after a reboot. An example on the table looks like this:
' Ethernet76     N/A     N/A     N/A     N/A     N/A     N/A     N/A     N/A'

How did you do it?
How did you verify/test it?
Ran the test case that no longer has the N/A value (pfc/test_pfc_pause_lossless_warm_reboot.py)

Any platform specific information?
No

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>