update autoneg setting to support 'off' when deploy-mg (#20030)
Summary:
Current 'autoneg' column in links.csv only support 'on'. If it's 'off' or other settings, it will default to platform.json behavior.
This PR add the support for 'off' settings.
For DUT which want to use the default behavior, it can leave that column empty, or use any other value, e.g. 'none'

Note: There will be a behavior change if user is using off in links.csv already for their DUT.
old behavior: the autoneg settings will be derived from platform.json, which chould be on or off or not defined.
new behavior: will be off always.
If users is using off already, they need to update their links.csv to leave autoneg field as empty if they want to use the default settings in platform.json.

What is the motivation for this PR?
update autoneg setting to support 'off'

How did you do it?
Check autoneg value for both on and off in minigraph

Signed-off-by: Yael Tzur <ytzur@nvidia.com>