Remove 7060CX sku from the kubesonic test not support list (#20003)
What is the motivation for this PR?
The 7060CX will run full image, so it's good to enable kubesonic test for these skus.

How did you do it?
Remove 7060CX sku from the kubesonic test not support list.

How did you verify/test it?
Run the test case on these sku's DUT to see if the kubesonic test will be skipped or not.

Signed-off-by: yenlu <yenlu@cisco.com>