[test_voq_disrupts.py] Ensure 'portA' and 'portB' are different in pick_ports (#19973)
Fixes #19971

What is the motivation for this PR?
When using pick_ports function in test_voq_ipfwd.py, it was possible that the function chose the same port for 'portA' and 'portB' when forced to choose 2 of the same types of ports. This was causing failures in test_voq_disrupts.py which used the function.

How did you do it?
Remove 'portA' from candidates for 'portB'.

How did you verify/test it?
Ran the failing test a few times to make sure it no longer fails

Signed-off-by: Javier Tan javiertan@microsoft.com
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>