conditional-mark: Evaluate github issues lazily
Instead of resolving status of all issues in a condition before calling `eval`,
defer checking each issue to the moment it is used. This way, if a condition
can be evaluated without needing the status of an issue, we can avoid the need
to make a request to the github api.

Implementation: instead of replacing issue url with 'True' or 'False', replace
it with a function call that will check the status of the issue. If the
condition short-circuits, the function call does not happen and no api request
is made.