Ignore router advertisements in PTF (#20113)
Ignore any router advertisements sent by a DUT, and don't set a default
route or an address based on it. This could happen if a T0 testbed with
radv running sends router advertisements on a VLAN interface, which may
result in the PTF container adding a default route on all of the VLAN
interfaces. This could result in some IPv6 test cases breaking.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>