Fix k8s node join time not enough issue (#20141)
What is the motivation for this PR?
It's an improvement for testbed vms75-t0-7050cx3-1. For this SKU, the kubelet needs more time(around 70s) to join the minikube cluster.

How did you do it?
Increased the wait time for joining node to cluster.

How did you verify/test it?
Run this test in testbed vms75-t0-7050cx3-1 to see if it passes.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>